Renamed GetPartiesForUser to GetPartyForUser and moved it from RegisterService to ContextRetrievalService.
Removed unused GetPartyForPartyId operation from RegisterService.
