+++ src/Altinn.AccessManagement.Core/Services/RegisterService.cs
@@ -29,50 +29,5 @@ namespace Altinn.AccessManagement.Core.Services
         {
             return _partyClient.LookupPartyBySSNOrOrgNo(organisationNumber);
         }
-
-        /// <inheritdoc/>
-        public Task<Party> GetPartyForPartyId(int partyId, int userId)
-        {
-            return _partyClient.GetPartyAsync(partyId);
-        }
-
-        /// <inheritdoc/>
-        public async Task<Party> GetPartiesForUser(int userId, int partyId)
-        {
-            List<Party> partyList = await _partyClient.GetPartiesForUserAsync(userId);
-
-            if (partyList.Count > 0)
-            {
-                foreach (Party party in partyList)
-                {
-                    if (party != null && party.PartyId == partyId)
-                    {
-                        return party;
-                    }
-                    else if (party != null && party.ChildParties != null && party.ChildParties.Count > 0)
-                    {
-                        return GetChildPartyIfMatch(party, partyId);
-                    }
-                }
-            }
-
-            return null;
-        }
-
-        private static Party GetChildPartyIfMatch(Party party, int partyId)
-        {
-            if (party.ChildParties != null)
-            {
-                foreach (Party childParty in party.ChildParties)
-                {
-                    if (childParty.PartyId == partyId)
-                    {
-                        return childParty;
-                    }
-                }
-            }
-
-            return null;
-        }
     }
 }
