+++ test/Altinn.AccessManagement.Tests/Controllers/LookupControllerTest.cs
@@ -10,9 +10,7 @@ using Altinn.AccessManagement.Tests.Mocks;
 using Altinn.AccessManagement.Tests.Util;
 using Altinn.AccessManagement.Tests.Utils;
 using Altinn.AccessManagement.Utilities;
-using Altinn.Platform.Register.Models;
 using AltinnCore.Authentication.JwtCookie;
-using AutoMapper;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.AspNetCore.TestHost;
 using Microsoft.Extensions.DependencyInjection;
